% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{download_cadastro_fundos}
\alias{download_cadastro_fundos}
\title{Retorna um csv com informações sobre os fundos de investimento}
\usage{
download_cadastro_fundos(download_path = NULL, verbose = TRUE)
}
\description{
Retorna o csv mais atual contendo informações sobre os fundos de investimento registrados na CVM
}
