% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{emissores_ativos_em_carteira}
\alias{emissores_ativos_em_carteira}
\title{Retorna uma tabela com informações sobre os ativos dos fundos de investimento}
\usage{
emissores_ativos_em_carteira(download_path = NULL, verbose = TRUE)
}
\description{
Retorna uma tabela contendo as mais atuais informações sobre os ativos de todos os fundos de investimento registrados na CVM
}
