% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{emissores_caracteristicas_fundos}
\alias{emissores_caracteristicas_fundos}
\title{Retorna uma tabela com informações sobre as características dos fundos}
\usage{
emissores_caracteristicas_fundos(download_path = NULL, verbose = TRUE)
}
\description{
Retorna uma tabela contendo as mais atuais informações sobre as características de todos os fundos de investimento registrados na CVM
}
