% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{emissores_informe_diario}
\alias{emissores_informe_diario}
\title{Retorna uma tabela contendo o informe diário}
\usage{
emissores_informe_diario(download_path = NULL, verbose = TRUE)
}
\description{
Retorna a tabela mais atual contendo o informe diário com informações agregadas sobre valor total de captação, valor total dos resgates, quantidade de cotistas, total da carteira e patrimônio líquido
}
