% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{retornar_quota_recente}
\alias{retornar_quota_recente}
\title{Retorna quota de fundos de investimento}
\usage{
retornar_quota_recente(CNPJ_BUSCA)
}
\arguments{
\item{CNPJ_BUSCA}{cpnj da do fundo}
}
\description{
Através do cnpj, retorna a quota de fundos de investimento a partir do último informe diário da CVM
}
